Handle state saving/restoring in top drawer.
