+++ library/src/net/simonvt/menudrawer/DraggableDrawer.java
@@ -5,7 +5,6 @@ import android.content.Context;
 import android.graphics.Canvas;
 import android.os.Build;
 import android.os.Bundle;
-import android.os.Parcel;
 import android.os.Parcelable;
 import android.util.AttributeSet;
 import android.view.MotionEvent;
@@ -588,69 +587,16 @@ public abstract class DraggableDrawer extends MenuDrawer {
      */
     protected abstract void drawIndicator(Canvas canvas, int offsetPixels);
 
-    public Parcelable saveState() {
-        Bundle state = new Bundle();
+    void saveState(Bundle state) {
         final boolean menuVisible = mDrawerState == STATE_OPEN || mDrawerState == STATE_OPENING;
         state.putBoolean(STATE_MENU_VISIBLE, menuVisible);
-        return state;
     }
 
     public void restoreState(Parcelable in) {
+        super.restoreState(in);
         Bundle state = (Bundle) in;
         final boolean menuOpen = state.getBoolean(STATE_MENU_VISIBLE);
         setOffsetPixels(menuOpen ? mMenuSize : 0);
         mDrawerState = menuOpen ? STATE_OPEN : STATE_CLOSED;
-    }
-
-    @Override
-    protected Parcelable onSaveInstanceState() {
-        Parcelable superState = super.onSaveInstanceState();
-
-        SavedState state = new SavedState(superState);
-        state.mMenuVisible = mDrawerState == STATE_OPEN || mDrawerState == STATE_OPENING;
-
-        return state;
-    }
-
-    @Override
-    protected void onRestoreInstanceState(Parcelable state) {
-        SavedState savedState = (SavedState) state;
-        super.onRestoreInstanceState(savedState.getSuperState());
-
-        setOffsetPixels(savedState.mMenuVisible ? mMenuSize : 0);
-        mDrawerState = savedState.mMenuVisible ? STATE_OPEN : STATE_CLOSED;
-    }
-
-    static class SavedState extends BaseSavedState {
-
-        boolean mMenuVisible;
-
-        public SavedState(Parcelable superState) {
-            super(superState);
-        }
-
-        public SavedState(Parcel in) {
-            super(in);
-            mMenuVisible = in.readInt() == 1;
-        }
-
-        @Override
-        public void writeToParcel(Parcel dest, int flags) {
-            super.writeToParcel(dest, flags);
-            dest.writeInt(mMenuVisible ? 1 : 0);
-        }
-
-        @SuppressWarnings("UnusedDeclaration")
-        public static final Creator<SavedState> CREATOR = new Creator<SavedState>() {
-            @Override
-            public SavedState createFromParcel(Parcel in) {
-                return new SavedState(in);
-            }
-
-            @Override
-            public SavedState[] newArray(int size) {
-                return new SavedState[size];
-            }
-        };
     }
 }
